+++ modulo6/rodada-cases-semana3/Case-cubo/Backend/src/business/UserBusiness.ts
@@ -1,0 +1,5 @@
+export class UserBusiness {
+  createUser =async () => {
+    
+  }
+}
