+++ modulo6/rodada-cases-semana3/Case-cubo/Backend/src/controller/UserController.ts
@@ -1,0 +1,5 @@
+export class UserController {
+  createUser = async (params:any) => {
+    
+  }
+}
