--- modulo6/rodada-cases-semana3/Case-cubo/Backend/src/controller/app.ts
@@ -1,19 +1,0 @@
-import express, {Express} from 'express';
-import dotenv from 'dotenv'
-import cors from 'cors'
-import { AddressInfo } from 'net'
-
-dotenv.config();
-
-export const app: Express = express();
-app.use(express.json())
-app.use(cors())
-
-const server = app.listen(process.env.PORT || 3003,()=>{
-    if(server){
-        const address = server.address() as AddressInfo;
-        console.log(`Servidor rodando na porta ${address.port}`)
-    }else{
-        console.log("server is not available")
-    }
-})
