+++ modulo6/rodada-cases-semana3/Case-cubo/Backend/src/routes/userRouter.ts
@@ -1,0 +1,11 @@
+import express from "express";
+import { UserBusiness } from "../business/UserBusiness";
+import { UserController } from "../controller/UserController";
+
+
+export const userRouter = express.Router();
+
+
+const userController = new UserController()
+
+userRouter.post("/create", userController.createUser);
